update modify packages script for Android and iOS

+++ src/Couchbase.Lite.Tests.Android/modify_packages.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
+if [ -f Couchbase.Lite.Tests.Android.Source.csproj ];
+then
 pushd `dirname $0`
 sed "s/[0-9].[0-9].[0-9]-b..../$1-$2/g" Couchbase.Lite.Tests.Android.csproj > tmp
 mv tmp Couchbase.Lite.Tests.Android.csproj
 sed "s/[0-9].[0-9].[0-9]-b..../$1-$2/g" Couchbase.Lite.Tests.Android.Source.csproj > tmp
 mv tmp Couchbase.Lite.Tests.Android.Source.csproj
 popd
+fi
